update storybook to work with node15
