+++ package.json
@@ -124,14 +124,14 @@
 		"@babel/plugin-transform-modules-commonjs": "^7.10.1",
 		"@babel/plugin-transform-runtime": "^7.11.5",
 		"@babel/preset-env": "^7.10.1",
-		"@storybook/addon-a11y": "^6.0.28",
-		"@storybook/addon-essentials": "^6.0.28",
-		"@storybook/addon-knobs": "^6.0.28",
-		"@storybook/addon-links": "^6.0.28",
-		"@storybook/addon-storysource": "^6.0.28",
-		"@storybook/addons": "^6.0.28",
+		"@storybook/addon-a11y": "^6.1.5",
+		"@storybook/addon-essentials": "^6.1.5",
+		"@storybook/addon-knobs": "^6.1.5",
+		"@storybook/addon-links": "^6.1.5",
+		"@storybook/addon-storysource": "^6.1.5",
+		"@storybook/addons": "^6.1.5",
 		"@storybook/storybook-deployer": "^2.8.7",
-		"@storybook/vue": "^6.0.28",
+		"@storybook/vue": "^6.1.5",
 		"@vue/test-utils": "^1.0.0-beta.29",
 		"app-manifest-loader": "^2.4.1",
 		"autoprefixer": "^9.8.6",
